test: add more cases in auction keeper
